引数をBook型に変更
